+++ src/ir/print.rs
@@ -1,3 +1,1 @@
-use super::IRProgram;
-
 pub struct IRPrinter;